refactor: move functions from path.js to utils.js
